requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-database-url==0.5.0
Django==4.2.16
django-allauth==65.2.0
django-anymail==13.0
django-bootstrap4==24.4
django-channels==0.7.0
django-ckeditor==6.7.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-dump-die==0.1.7
django-environ==0.8.1
django-filter==25.1
django-htmx==1.21.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
docopt==0.6.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
num2words==0.5.14
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pi==0.1.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.8.0
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
pytz==2021.3
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
toml==0.10.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.3
whitenoise==5.3.0
zope.interface==7.1.1